+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.14] - 2022-03-18
+
+### Updated
+
+- Accounts API to include assign and get countries, with examples
+- Updated typescript types
+- Updated documentation
+
 ## [0.1.13] - 2022-03-17
 
 ### Added